RISC-V 64bit: shared kernel files modifications

+++ include/kernel/boot.h
@@ -78,7 +78,7 @@ create_frames_of_region(
     cap_t    pd_cap,
     region_t reg,
     bool_t   do_map,
-    int32_t  pv_offset
+    int64_t  pv_offset
 );
 
 cap_t
